Add new tab for libdotdesktop file interpretation.
This one looks better than the one that's currently being used, and has a textbox for output rather than using labels. It should also be able to be resized if needed, and the controls will move around appropriately. The table layout panel still needs to have the rows properly locked in place for size at the top, though.