Properly open the program after getting...
a URL from the user, if the .desktop file can use one.

Also remove space to the right of the special fields when seeing if they exist in the Exec key since they probably wouldn't have space after them if they're the last in the file. The output for what program is being launched has also been moved to just before the Process.Start code for that program.