Maybe, possibly allowing args to be passed from...
a desktop entry file.

Right now, the cleaned exec key is copied into an original version for later use in the arg variable, as well as into a string that gets split at the quotation marks. Once it's copied, the temporary one to be split is split at the quotation marks and it has whitespace trimmed off. Afterward, the arg variable takes the original exec key copy and replaces the quotation marks surrounding the new exec key as well as the exec key itself. For testing purposes, these keys are shown in messageboxes.