Use a question mark for splitter instead of semicolon...
as the question mark is not allowed in NTFS filenames and this may be more likely to not have a problem. This could be a configurable option so the user can change it to a character that doesn't conflict with anything else if they want to or if they run into issues. Maybe have that be an advanced option.