Allow editing the file list before launching.
This would be useful if the user wants to open a file in something like nano. Not quite perfect yet as it seems to have issues with multiple files in the list. Maybe what should be done instead is a window that allows switching between Windows-style and Linux-style paths as well as a manual editing option.